SB-38 add filtering methods

+++ src/main/java/com/tricast/repositories/entities/League.java
@@ -1,24 +1,30 @@
 package com.tricast.repositories.entities;
 
 import java.io.Serializable;
-
 import java.util.HashSet;
 import java.util.Set;
 
-import javax.persistence.*;
+import javax.persistence.CascadeType;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.FetchType;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.JoinTable;
+import javax.persistence.ManyToMany;
+import javax.persistence.Table;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
 @Entity
 @Table(name = "LEAGUES")
 public class League implements Serializable {
-    
-	/**
-	 * 
-	 */
-	private static final long serialVersionUID = 2318290506163896020L;
-	
-	@Id
+
+    private static final long serialVersionUID = 2318290506163896020L;
+
+    @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
@@ -26,46 +32,42 @@ public class League implements Serializable {
     private String description;
 
     @Column(name = "sportid")
-    private Integer sportId;
+    private Long sportId;
 
-    @ManyToMany(fetch = FetchType.LAZY ,cascade = {CascadeType.ALL})
-    @JoinTable(name = "LEAGUECOMPETITORMAP",
-        joinColumns = 		 @JoinColumn(name="leagueid", referencedColumnName="id"),
-        inverseJoinColumns = @JoinColumn(name="competitorid", referencedColumnName="id")
-        )
-	@JsonIgnoreProperties("leagues")
+    @ManyToMany(fetch = FetchType.LAZY, cascade = { CascadeType.ALL })
+    @JoinTable(name = "LEAGUECOMPETITORMAP", joinColumns = @JoinColumn(name = "leagueid", referencedColumnName = "id"), inverseJoinColumns = @JoinColumn(name = "competitorid", referencedColumnName = "id"))
+    @JsonIgnoreProperties("leagues")
     public Set<Competitor> competitors = new HashSet<>();
 
-    
-	public Set<Competitor> getCompetitors() {
-		return competitors;
-	}
+    public Set<Competitor> getCompetitors() {
+        return competitors;
+    }
 
-	public void setCompetitors(Set<Competitor> competitors) {
-		this.competitors = competitors;
-	}
+    public void setCompetitors(Set<Competitor> competitors) {
+        this.competitors = competitors;
+    }
 
-	public Long getId() {
-		return id;
-	}
+    public Long getId() {
+        return id;
+    }
 
-	public void setId(Long id) {
-		this.id = id;
-	}
+    public void setId(Long id) {
+        this.id = id;
+    }
 
-	public String getDescription() {
-		return description;
-	}
+    public String getDescription() {
+        return description;
+    }
 
-	public void setDescription(String description) {
-		this.description = description;
-	}
+    public void setDescription(String description) {
+        this.description = description;
+    }
 
-	public Integer getSportId() {
-		return sportId;
-	}
+    public Long getSportId() {
+        return sportId;
+    }
 
-	public void setSportId(Integer sportId) {
-		this.sportId = sportId;
-	}
+    public void setSportId(Long sportId) {
+        this.sportId = sportId;
+    }
 }
